@@ -1,3 +1,5 @@
 module.exports = {
-  presets: ["@vue/cli-plugin-babel/preset"]
-};
+  presets: [
+    '@vue/cli-plugin-babel/preset'
+  ]
+}
